WIP: The spherical diffuse interface does not yield the expected results In particular, with a cavity inside a bubble, the result is identical to vacuum (no dielectric effect) I suspect this is connected to the removal of the L=0 component from the image part of the Green's function.
Additional note: the integration of the radial part of the green's function can be significantly simplified by making heavilly use of the asymptotic solutions
Moreover I believe that an additional change of variable would make the integration even faster, because, in a logarithmic scale all components are straight lines, except for a small deviation in the interfacial region. Maybe solving directly for Gimg could indeed help.